move executors into main file
